feat: Export Project

+++ resources/views/livewire/project/project-item.blade.php
@@ -54,17 +54,17 @@
     </td>
 
     <td>
-        @can('update:project')
-            <a href="{{ route('project.edit', ['id' => $project->id]) }}"
-                class="btn btn-primary btn-sm waves-effect waves-light"><i class="bx bx-pencil me-1"></i> Edit</a>
-        @endcan
-        @can('delete:project')
-            <li class="list-inline-item">
+        @if ($project->employee_id == Auth::user()->employee->id)
+            @can('update:project')
+                <a href="{{ route('project.edit', ['id' => $project->id]) }}"
+                    class="btn btn-primary btn-sm waves-effect waves-light"><i class="bx bx-pencil me-1"></i> Edit</a>
+            @endcan
+            @can('delete:project')
                 <button type="button" wire:click="deleteConfirm()" class="btn btn-danger btn-sm waves-effect waves-light">
                     <i class="mdi mdi-delete me-1"></i> Delete
                 </button>
-            </li>
-        @endcan
+            @endcan
+        @endif
         <a href="{{ route('project.detail', ['id' => $project->id]) }}"
             class="btn btn-primary btn-sm waves-effect waves-light"><i class="mdi mdi-eye me-1"></i> Detail</a>
     </td>
